Update resolved reference in poetry.lock
The resolved reference for the 'BencherServer' package source in poetry.lock file has been updated. This aids in enhancing the efficiency of the benchmark setup and cloning process inside the container. It also addresses unnecessary duplication and ordering issues during the git cloning operation. The 'mopta08' benchmark was furthermore included in the client configuration.